# Conflicts:
#	features/user/onboarding/build.gradle
#	features/user/onboarding/src/main/java/com/tokopedia/onboarding/view/fragment/OnboardingFragment.kt

+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/search/presentation/viewmodel/HotelSearchResultViewModel.kt
@@ -34,7 +34,7 @@ class HotelSearchResultViewModel @Inject constructor(
 
     var isFilter = false
 
-    fun initSearchParam(destinationID: Int, type: String, latitude: Float, longitude: Float,
+    fun initSearchParam(destinationID: Long, type: String, latitude: Float, longitude: Float,
                         checkIn: String, checkOut: String, totalRoom: Int, totalAdult: Int) {
         if (type == TYPE_CITY)
             searchParam.location.cityID = destinationID
